+++ pprof-parser/src/main/java/io/jafar/pprof/api/PprofProfile.java
@@ -1,4 +1,4 @@
-package io.jafar.pprof.shell;
+package io.jafar.pprof.api;
 
 import java.util.List;